refactor: change Check class to Type-Extensions
pick one word per concept
