+++ source/test/F0.CodeAnalysis.Testing/CodeAnalysis/CodeRefactorings/CodeRefactoringVerifier.cs
@@ -3,7 +3,6 @@ using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 using F0.Testing.Extensions;
-using F0.Testing.Shared;
 using Microsoft.CodeAnalysis.CodeRefactorings;
 using Microsoft.CodeAnalysis.CSharp;
 
@@ -16,10 +15,10 @@ namespace F0.Testing.CodeAnalysis.CodeRefactorings
 		{
 			var type = typeof(TCodeRefactoring);
 
-			Check.Accessibility(type);
-			Check.NonInheritable(type);
-			Check.ExportCodeRefactoringProviderAttribute(type);
-			Check.SharedAttribute(type);
+			type.VerifyAccessibility();
+			type.VerifyNonInheritable();
+			type.VerifyExportCodeRefactoringProviderAttribute();
+			type.VerifySharedAttribute();
 		}
 
 		public Task NoOpAsync(string code)
